+++ src/Resettable.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Duon\Container;
+
+interface Resettable
+{
+	public function reset(): void;
+}